+++ src/apis/logging_kafka_api.rs
@@ -54,7 +54,7 @@ pub struct CreateLogKafkaParams {
     pub user: Option<String>,
     /// SASL password.
     pub password: Option<String>,
-    pub use_tls: Option<crate::models::LoggingUseTls>
+    pub use_tls: Option<crate::models::LoggingUseTlsString>
 }
 
 /// struct for passing parameters to the method [`delete_log_kafka`]