+++ src/models/logging_kafka_response.rs
@@ -78,7 +78,7 @@ pub struct LoggingKafkaResponse {
     #[serde(rename = "password", skip_serializing_if = "Option::is_none")]
     pub password: Option<String>,
     #[serde(rename = "use_tls", skip_serializing_if = "Option::is_none")]
-    pub use_tls: Option<crate::models::LoggingUseTls>,
+    pub use_tls: Option<crate::models::LoggingUseTlsString>,
 }
 
 impl LoggingKafkaResponse {