+++ src/models/logging_splunk_additional.rs
@@ -17,7 +17,7 @@ pub struct LoggingSplunkAdditional {
     #[serde(rename = "token", skip_serializing_if = "Option::is_none")]
     pub token: Option<String>,
     #[serde(rename = "use_tls", skip_serializing_if = "Option::is_none")]
-    pub use_tls: Option<crate::models::LoggingUseTls>,
+    pub use_tls: Option<crate::models::LoggingUseTlsString>,
 }
 
 impl LoggingSplunkAdditional {