+++ src/models/logging_syslog_additional.rs
@@ -22,7 +22,7 @@ pub struct LoggingSyslogAdditional {
     #[serde(rename = "token", skip_serializing_if = "Option::is_none")]
     pub token: Option<String>,
     #[serde(rename = "use_tls", skip_serializing_if = "Option::is_none")]
-    pub use_tls: Option<crate::models::LoggingUseTls>,
+    pub use_tls: Option<crate::models::LoggingUseTlsString>,
 }
 
 impl LoggingSyslogAdditional {